backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
openai==1.3.8
pymupdf==1.23.8
pillow==10.1.0
pytesseract==0.3.10
pdf2image==1.16.3
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0
